update cricket service functions
