+++ sports-scores/src/app/sports/cricket/series/[ccd]/[scd]/matches/page.tsx
@@ -1,20 +1,17 @@
 import CricketFixtureList from "@/components/cricket/CricketFixtureList";
 import Placeholder from "@/components/misc/Placeholder";
-import { fetchCricketSeriesMatches } from "@/endpoints/cricket.api";
-import { mapCricketSeriesMatches } from "@/lib/dataMapping";
+import { cricketSeriesDetails } from "@/services/cricket.service";
 
 export default async function Page(props: {
   params: Promise<{ ccd: string; scd: string }>;
 }) {
   const params = await props.params;
 
-  let rawMatches = await fetchCricketSeriesMatches(params.ccd, params.scd);
+  let matches = await cricketSeriesDetails(params.ccd, params.scd);
 
-  if (rawMatches == undefined || rawMatches.Stages == undefined) {
+  if (matches == null) {
     return <Placeholder>An error has ocurred</Placeholder>;
   }
 
-  let fixtures = mapCricketSeriesMatches(rawMatches);
-
-  return <CricketFixtureList data={fixtures} />;
+  return <CricketFixtureList data={matches} />;
 }
